requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
datasets==2.10.1
deepgram-sdk==3.3.1
dill==0.3.6
distro==1.9.0
filelock==3.15.4
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
google==3.0.0
greenlet==3.0.3
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain
langchain-community
langchain-core
langchain-openai
langchain-experimental
langchain-pinecone
langsmith
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
multiprocess==0.70.14
mypy-extensions==1.0.0
numpy==1.26.4
openai
orjson==3.10.6
packaging==23.2
pandas==2.2.2
pinecone-client
protobuf==5.27.2
psycopg2==2.9.9
pyarrow==16.1.0
pydantic==2.8.2
pydantic_core==2.20.1
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
responses==0.18.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
tenacity==8.4.2
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
xxhash==3.4.1
yarl==1.9.4
pytest
gunicorn
selenium
webdriver_manager